+++ apps/worker/job-handlers.ts
@@ -133,15 +133,6 @@ const setupJobHandlers = async (name: string, chainId: string, res: any, next: N
                 next,
             );
             break;
-        case 'update-liquidity-for-active-pools':
-            await runIfNotAlreadyRunning(
-                name,
-                chainId,
-                () => PoolController().updateLiquidityValuesForActivePools(chain),
-                res,
-                next,
-            );
-            break;
         case 'update-liquidity-for-inactive-pools':
             await runIfNotAlreadyRunning(
                 name,
@@ -211,9 +202,6 @@ const setupJobHandlers = async (name: string, chainId: string, res: any, next: N
                 res,
                 next,
             );
-            break;
-        case 'update-lifetime-values-for-all-pools':
-            await runIfNotAlreadyRunning(name, chainId, () => poolService.updateLifetimeValuesForAllPools(), res, next);
             break;
         case 'feed-data-to-datastudio':
             await runIfNotAlreadyRunning(
@@ -378,16 +366,6 @@ const setupJobHandlers = async (name: string, chainId: string, res: any, next: N
                 next,
             );
             break;
-        // TODO
-        // case 'update-lifetime-values-for-all-pools-v3':
-        //     await runIfNotAlreadyRunning(
-        //         name,
-        //         chainId,
-        //         () => poolService.updateLifetimeValuesForAllPoolsV3(),
-        //         res,
-        //         next,
-        //     );
-        //     break;
         // COW AMM
         case 'sync-cow-amm-pools':
             await runIfNotAlreadyRunning(name, chainId, () => CowAmmController().syncPools(chain), res, next);
@@ -400,9 +378,6 @@ const setupJobHandlers = async (name: string, chainId: string, res: any, next: N
             break;
         case 'sync-cow-amm-snapshots':
             await runIfNotAlreadyRunning(name, chainId, () => CowAmmController().syncSnapshots(chain), res, next);
-            break;
-        case 'update-cow-amm-volume-and-fees':
-            await runIfNotAlreadyRunning(name, chainId, () => CowAmmController().updateVolumeAndFees(chain), res, next);
             break;
         case 'sync-categories':
             await runIfNotAlreadyRunning(name, chainId, () => ContentController().syncCategories(), res, next);