+++ modules/network/sepolia.ts
@@ -32,22 +32,16 @@ export const sepoliaNetworkConfig: NetworkConfig = {
         ),
     },
     /*
-    For sub-minute jobs we set the alarmEvaluationPeriod and alarmDatapointsToAlarm to 1 instead of the default 3. 
+    For sub-minute jobs we set the alarmEvaluationPeriod and alarmDatapointsToAlarm to 1 instead of the default 3.
     This is needed because the minimum alarm period is 1 minute and we want the alarm to trigger already after 1 minute instead of 3.
 
-    For every 1 days jobs we set the alarmEvaluationPeriod and alarmDatapointsToAlarm to 1 instead of the default 3. 
+    For every 1 days jobs we set the alarmEvaluationPeriod and alarmDatapointsToAlarm to 1 instead of the default 3.
     This is needed because the maximum alarm evaluation period is 1 day (period * evaluationPeriod).
     */
     workerJobs: [
         {
             name: 'update-liquidity-for-inactive-pools',
-            interval: every(1, 'days'),
-            alarmEvaluationPeriod: 1,
-            alarmDatapointsToAlarm: 1,
-        },
-        {
-            name: 'update-liquidity-for-active-pools',
-            interval: every(2, 'minutes'),
+            interval: every(10, 'minutes'),
         },
         {
             name: 'update-pool-apr',
@@ -76,10 +70,6 @@ export const sepoliaNetworkConfig: NetworkConfig = {
         {
             name: 'sync-snapshots-v2',
             interval: every(90, 'minutes'),
-        },
-        {
-            name: 'update-lifetime-values-for-all-pools',
-            interval: every(50, 'minutes'),
         },
         {
             name: 'sync-changed-pools',
@@ -160,10 +150,6 @@ export const sepoliaNetworkConfig: NetworkConfig = {
             interval: every(1, 'minutes'),
         },
         { name: 'sync-cow-amm-snapshots', interval: every(90, 'minutes') },
-        {
-            name: 'update-cow-amm-volume-and-fees',
-            interval: every(20, 'minutes'),
-        },
         {
             name: 'sync-erc4626-unwrap-rate',
             interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(60, 'minutes') : every(20, 'minutes'),