+++ modules/sor/lib/sor.ts
@@ -16,7 +16,15 @@ import {
 import { BasePool } from './poolsV2/basePool';
 import { SorSwapOptions } from './types';
 import { PathWithAmount } from './path';
-import { BufferPool, Gyro2CLPPool, GyroECLPPool, ReClammPool, StablePoolV3, WeightedPoolV3 } from './poolsV3';
+import {
+    BufferPool,
+    Gyro2CLPPool,
+    GyroECLPPool,
+    QuantAmmPool,
+    ReClammPool,
+    StablePoolV3,
+    WeightedPoolV3,
+} from './poolsV3';
 import { LiquidityBootstrappingPoolV3 } from './poolsV3/liquidityBootstrapping/liquidityBootstrapping';
 import { BufferPoolData } from '../utils/data';
 
@@ -33,7 +41,7 @@ export class SOR {
     ): Promise<PathWithAmount[] | null> {
         const checkedSwapAmount = checkInputs(tokenIn, tokenOut, swapKind, swapAmountEvm);
 
-        // get current block timestamp for ReClamm math
+        // get current block timestamp if not provided
         const currentTimestamp = swapOptions?.currentTimestamp ?? BigInt(Date.now()) / 1000n;
 
         const basePools: BasePool[] = [];
@@ -48,7 +56,6 @@ export class SOR {
             }
             switch (prismaPool.type) {
                 case 'WEIGHTED':
-                case 'QUANT_AMM_WEIGHTED':
                     {
                         if (prismaPool.protocolVersion === 2) {
                             basePools.push(WeightedPool.fromPrismaPool(prismaPool));
@@ -111,6 +118,9 @@ export class SOR {
                     break;
                 case 'RECLAMM':
                     basePools.push(ReClammPool.fromPrismaPool(prismaPool, currentTimestamp));
+                    break;
+                case 'QUANT_AMM_WEIGHTED':
+                    basePools.push(QuantAmmPool.fromPrismaPool(prismaPool, currentTimestamp));
                     break;
                 default:
                     console.log('Unsupported pool type');