+++ prisma/migrations/20250522215153_pool_token_update_index/migration.sql
@@ -1,0 +1,2 @@
+-- CreateIndex
+CREATE INDEX "PrismaPoolToken_id_chain_idx" ON "PrismaPoolToken"("id", "chain");