+++ test/testData/read/mapping.ts
@@ -8,6 +8,7 @@ import { getDecimalsFromScalingFactor } from '../../utils';
 import {
     BufferPool,
     GyroEPool,
+    QuantAmmPool,
     ReClammPool,
     StablePool,
     WeightedPool,
@@ -392,6 +393,86 @@ export function mapReClammPoolStateToPrismaPool(
             priceRatioUpdateStartTime: Number(poolState.priceRatioUpdateStartTime),
             dailyPriceShiftBase: formatEther(poolState.dailyPriceShiftBase),
             startFourthRootPriceRatio: formatEther(poolState.startFourthRootPriceRatio),
+        })
+        .build({
+            address: poolState.poolAddress,
+            protocolVersion,
+            tokens,
+            dynamicData: {
+                swapFee: formatEther(poolState.swapFee),
+                aggregateSwapFee: formatEther(poolState.aggregateSwapFee),
+                totalShares: formatEther(poolState.totalSupply),
+            },
+            hook: {
+                ...poolState.hook,
+                dynamicData: hookDynamicData,
+            },
+            chain: chainIdToChain[chainId],
+        });
+    return prismaPool;
+}
+
+export function mapQuantAmmPoolStateToPrismaPool(
+    poolState: QuantAmmPool,
+    chainId: number,
+    protocolVersion: number,
+    bufferPools: BufferPool[],
+): PrismaPoolAndHookWithDynamic {
+    const decimals = poolState.scalingFactors.map((scalingFactor: bigint) =>
+        getDecimalsFromScalingFactor(scalingFactor),
+    );
+
+    const poolTokens = poolState.tokens.map(
+        (token: string, i: number) => new Token(chainId, token as Address, decimals[i]),
+    );
+
+    const tokenAmounts = poolTokens.map((token: Token, i: number) =>
+        PoolTokenWithRate.fromScale18AmountWithRate(
+            token,
+            poolState.balancesLiveScaled18[i],
+            poolState.tokenRates[i],
+            i,
+        ),
+    );
+
+    // map tokenIn and tokenOut to prisma tokens using prisma token factory
+    const tokens = poolState.tokens.map((_token, i) => {
+        const bufferPool = bufferPools.find((bufferPool) =>
+            isSameAddress(_token as Address, bufferPool.poolAddress as Address),
+        );
+        const token = bufferPool
+            ? {
+                  decimals: decimals[i],
+                  unwrapRate: formatUnits(bufferPool.rate, 18 - decimals[i] + bufferPool.decimals[1]),
+                  underlyingTokenAddress: bufferPool.tokens[1],
+                  chain: chainIdToChain[chainId],
+              }
+            : { decimals: decimals[i], chain: chainIdToChain[chainId] };
+
+        return prismaPoolTokenFactory.build({
+            address: _token as Address,
+            balance: formatUnits(tokenAmounts[i].amount, decimals[i]),
+            index: i,
+            priceRate: formatEther(poolState.tokenRates[i]),
+            token,
+            chain: chainIdToChain[chainId],
+        });
+    });
+
+    // transform hook dynamicData values to bigInt and then apply formatEther to them
+    const _hookDynamicData = poolState.hook?.dynamicData;
+    const hookDynamicData = _hookDynamicData
+        ? Object.fromEntries(Object.entries(_hookDynamicData).map(([key, value]) => [key, formatEther(BigInt(value))]))
+        : undefined;
+
+    // map pool state to prisma pool using prisma pool factory
+    const prismaPool = prismaPoolFactory
+        .quantAmm({
+            firstFourWeightsAndMultipliers: poolState.firstFourWeightsAndMultipliers.map((w) => formatEther(w)),
+            secondFourWeightsAndMultipliers: poolState.secondFourWeightsAndMultipliers.map((w) => formatEther(w)),
+            maxTradeSizeRatio: formatEther(poolState.maxTradeSizeRatio),
+            lastUpdateIntervalTime: String(poolState.lastUpdateTime),
+            lastInterpolationTimePossible: String(poolState.lastInteropTime),
         })
         .build({
             address: poolState.poolAddress,