+++ test/testData/read/readTestData.ts
@@ -1,26 +1,28 @@
 import {
     BufferState,
     GyroECLPState,
+    QuantAmmState,
     ReClammState,
     StableState,
     WeightedState,
     BasePoolState,
     WeightedImmutable,
 } from '@balancer-labs/balancer-maths';
+import { Address } from '@balancer/sdk';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 
 import { HookData, PrismaPoolAndHookWithDynamic } from '../../../prisma/prisma-types';
+import { BufferPoolData } from '../../../modules/sor/utils/data';
+import { TransformBigintToString } from '../types';
 import {
     mapGyroPoolStateToPrismaPool,
+    mapQuantAmmPoolStateToPrismaPool,
     mapReClammPoolStateToPrismaPool,
     mapStablePoolStateToPrismaPool,
     mapWeightedPoolStateToPrismaPool,
     mapLiquidityBootstrappingPoolStateToPrismaPool,
 } from './mapping';
-import { Address } from '@balancer/sdk';
-import { BufferPoolData } from '../../../modules/sor/utils/data';
-import { start } from 'node:repl';
 
 type PoolBase = {
     poolAddress: string;
@@ -59,13 +61,16 @@ export type LiquidityBootstrappingPool = PoolBase & LiquidityBootstrappingState;
 
 export type ReClammPool = PoolBase & ReClammState;
 
+export type QuantAmmPool = PoolBase & QuantAmmState;
+
 export type SupportedPools =
     | WeightedPool
     | StablePool
     | BufferPool
     | GyroEPool
     | LiquidityBootstrappingPool
-    | ReClammPool;
+    | ReClammPool
+    | QuantAmmPool;
 
 type SwapPath = {
     swapKind: number;
@@ -133,7 +138,7 @@ export function readTestData(): TestData {
                     amountRaw: BigInt(jsonData.swapPath.amountRaw),
                     outputRaw: BigInt(jsonData.swapPath.outputRaw),
                     test: file,
-                    currentTimestamp,
+                    currentTimestamp: currentTimestampString ? BigInt(currentTimestampString) : undefined,
                     chainId: jsonData.test.chainId,
                 });
             } catch (error) {
@@ -144,10 +149,6 @@ export function readTestData(): TestData {
 
     return testData;
 }
-
-type TransformBigintToString<T> = {
-    [K in keyof T]: T[K] extends bigint ? string : T[K] extends bigint[] ? string[] : T[K];
-};
 
 function mapPools(pools: TransformBigintToString<SupportedPools>[]): PrismaPoolAndHookWithDynamic[] {
     const bufferPools: BufferPool[] = pools
@@ -260,6 +261,25 @@ function mapPools(pools: TransformBigintToString<SupportedPools>[]): PrismaPoolA
                 currentTimestamp: BigInt(pool.currentTimestamp),
             };
             prismaPools.push(mapReClammPoolStateToPrismaPool(reClammPool, Number(pool.chainId), 3, bufferPools));
+        } else if (pool.poolType === 'QUANT_AMM_WEIGHTED') {
+            const quantAmmPool = {
+                ...pool,
+                scalingFactors: pool.scalingFactors.map((sf) => BigInt(sf)),
+                swapFee: BigInt(pool.swapFee),
+                balancesLiveScaled18: pool.balancesLiveScaled18.map((b) => BigInt(b)),
+                tokenRates: pool.tokenRates.map((r) => BigInt(r)),
+                totalSupply: BigInt(pool.totalSupply),
+                aggregateSwapFee: BigInt(pool.aggregateSwapFee ?? '0'),
+                supportsUnbalancedLiquidity:
+                    pool.supportsUnbalancedLiquidity === undefined ? true : pool.supportsUnbalancedLiquidity,
+                firstFourWeightsAndMultipliers: pool.firstFourWeightsAndMultipliers.map((w) => BigInt(w)),
+                secondFourWeightsAndMultipliers: pool.secondFourWeightsAndMultipliers.map((m) => BigInt(m)),
+                maxTradeSizeRatio: BigInt(pool.maxTradeSizeRatio),
+                lastUpdateTime: BigInt(pool.lastUpdateTime),
+                lastInteropTime: BigInt(pool.lastInteropTime),
+                currentTimestamp: BigInt(pool.currentTimestamp),
+            };
+            prismaPools.push(mapQuantAmmPoolStateToPrismaPool(quantAmmPool, Number(pool.chainId), 3, bufferPools));
         }
     }
     return prismaPools;